REFACTOR. `browserSync.stream() is not required
Steps:
- Remove the `.pipe(browserSync.stream())` function in the `cssCompile` task, it is not required if you already have the server reload correctly configured in the `watch`.

Reason:
- The `.pipe(browserSync.stream())` function is used to reflect changes made to `CSS` files in the browser without having to reload the entire page. When a modification is made to a `CSS` file and saved, `Gulp` runs the `cssCompile` task, which compiles and minifies the `CSS` code and generates the `styles.min.css` file in the `dist/css/` directory.
Without the `.pipe(browserSync.stream())` line, even if the `CSS` file has been updated, the changes will not be automatically reflected in the browser. The browser will continue to display the previous version of the `CSS` file until you manually reload the page.
However, since you already have a `watch` configuration with `browserSync` in the `watch` task.
The `reload` function will be triggered whenever there is a change to the `HTML`, `CSS` or `JS` files inside the `dist/` directory. This will cause the browser to automatically reload to show the recent changes.
So, in your case, the `.pipe(browserSync.stream())` line in the `cssCompile` task is redundant and not necessary, since the change in the `CSS` files will already trigger the reloading of the browser through the watch configuration in the `watch` task. You can safely remove that line, and the behavior will remain the same: the `CSS` changes will be automatically reflected in the browser when you save the files.